requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.1.0
pydantic>=2.10.0
httpx>=0.28.0
anthropic>=0.52.0
openai>=1.82.0
google-generativeai>=0.8.0
openpyxl>=3.1.0
fastmcp>=2.0.0
